% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R
\name{ab2c}
\alias{ab2c}
\title{Calculate c in Focus (c, 0)}
\usage{
ab2c(a = 1, ab)
}
\arguments{
\item{a}{Semi-major (Ellipse) or Radium (Ring).}

\item{ab}{Semi-major over semi-minor. ab=1 for a Ring.}
}
\value{
c in Focus (c, 0)
}
\description{
Calculate c in Focus (c, 0)
}
\examples{
x1=PCS2CCS(a=10, ab=1.5)
x2=PCS2CCS(a=9, ab=1.2)
c1 = ab2c(a=10, ab=1.5)
c2 = ab2c(a=9, ab=1.2)
plot(x1, type='n', xlim=c(-10,10), ylim=c(-10,10), asp=1)
abline(h=0, v=0, asp=1, lty=2)
lines(x1, col=2);
points(c1, 0, col=2)
lines(x2, col=3);
points(c2, 0, col=3)

}
