% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3d.R
\name{Arrow3D}
\alias{Arrow3D}
\title{Plot 3D Arrow axis.
\code{Arrow3D}}
\usage{
Arrow3D(len = 10, orig = c(0, 0, 0), cols = c(2:4), ...)
}
\arguments{
\item{len}{Length of the arrow.}

\item{orig}{Origin of the axis.}

\item{cols}{Colors of axis.}

\item{...}{More options of arrow3d().}
}
\description{
Plot 3D Arrow axis.
\code{Arrow3D}
}
