% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R
\name{d2r}
\alias{d2r}
\title{Degree to Radian}
\usage{
d2r(x)
}
\arguments{
\item{x}{Degree}
}
\value{
Radian
}
\description{
Degree to Radian
}
\examples{
r = (1:100)/100 * 4 * pi
d = r2d(r)
rr = d2r(d)
plot(d, sin(rr));
abline(h=0 )
abline(v = 360)
}
