% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MathOrbit.R
\name{orbit.parameter}
\alias{orbit.parameter}
\title{Give the orbit the parameter}
\usage{
orbit.parameter(a, b = NULL, ab = NULL)
}
\arguments{
\item{a}{Semi-major axis}

\item{b}{Semi-minor axis}

\item{ab}{Semi-major over semi-minor. ab=1 for a Ring.}
}
\description{
Give the orbit the parameter
}
\examples{

orbit.parameter(a=1, b=1.5)

}
