% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2d.R
\name{plotfan}
\alias{plotfan}
\title{Plot a fan}
\usage{
plotfan(theta, a = 1, ab = 1, orig = c(0, 0), fun = graphics::plot,
  ...)
}
\arguments{
\item{theta}{Angle in polar coordinate system}

\item{a}{Radius of start and end points of the arrow.}

\item{ab}{Semi-major over semi-minor. ab=1 for a Ring.}

\item{orig}{Origin}

\item{fun}{Plot function. default = plot}

\item{...}{More options in plot function}
}
\description{
Plot a fan
}
\examples{

n=50
theta =0:n
plotpcs(theta = theta, a=1, ab=1, asp=1)
plotfan(theta = theta * 10, a=1, ab=1, fun=polygon)

}
