% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R
\name{r2d}
\alias{r2d}
\title{Radian to degree}
\usage{
r2d(x)
}
\arguments{
\item{x}{Radian}
}
\value{
Degree
}
\description{
Radian to degree
}
\examples{
r = (1:100)/100 * 4 * pi
d = r2d(r)
rr = d2r(d)
plot(d, sin(rr));
abline(h=0 )
abline(v = 360)
}
