% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R
\name{rotate}
\alias{rotate}
\title{Roate a (x,y) with an Angle.}
\usage{
rotate(x, theta, orig = c(0, 0))
}
\arguments{
\item{x}{Coordinate (x,y)}

\item{theta}{An angle (degree)}

\item{orig}{Pivot of rotation}
}
\value{
Rotated (x,y)
}
\description{
Roate a (x,y) with an Angle.
}
\examples{
plot(PCS2CCS(a=2), asp=1)
grid()
x=rbind(c(1,1))
for(i in 1:12){
  points(rotate(x, 30*i, orig=c(1,0)))
}
}
