% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{SpaceObject-class}
\alias{SpaceObject-class}
\alias{SpaceObject}
\title{Class of planet
\code{SpaceObject}}
\value{
Class of SpaceObject
}
\description{
Class of planet
\code{SpaceObject}
}
\section{Slots}{

\describe{
\item{\code{shape}}{Ploting function of the shape}

\item{\code{radius}}{Radius for sphere}

\item{\code{Period.Rotate}}{data.frame 1*3 c(Period.Rotate, Period.Orbit, Period.Synodic)}
}}

