% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{SpaceOrbit-class}
\alias{SpaceOrbit-class}
\alias{SpaceOrbit}
\title{Class of Orbit
\code{Orbit}}
\value{
Class of SpaceOrbit
}
\description{
Class of Orbit
\code{Orbit}
}
\section{Slots}{

\describe{
\item{\code{ab}}{Shape of the object, ab=1 Sphere, ab!=1 Ellipsoid}

\item{\code{e}}{eccentric of the orbit}

\item{\code{radius}}{Radius for sphere (ab=1), or Semi-major axis for Ellipsoid (ab!=1)}

\item{\code{period}}{data.frame 1*3 c(Period.Rotate, Period.Orbit, Period.Synodic)}

\item{\code{Inclination}}{Inclination.}

\item{\code{CenterObject}}{Central Object.}
}}

